feat(map-calibration): lift GrayImage + pure image ops into Mithril.MapCalibration (#914)
Lifts the BCL-only image-buffer primitives (GrayImage + Detection record)
and the three pure pixel ops (Resize/Downsample/Crop) out of the offline
tools into src/Mithril.MapCalibration/Detection/ (namespace
Mithril.MapCalibration.Detection). The System.Drawing load/save/draw stays in
ImageIo.cs (tools-only). Tools.Common, the probe, the screenshot calibrator
and the study tool now consume the lifted types/ops and green-build (§D6
dedup guard).
